+++ pkg/azureupdate/validate_azureconfig_test.go
@@ -2,7 +2,6 @@ package azureupdate
 
 import (
 	"context"
-	"encoding/json"
 	"testing"
 
 	providerv1alpha1 "github.com/giantswarm/apiextensions/v3/pkg/apis/provider/v1alpha1"
@@ -10,12 +9,11 @@ import (
 
 	"github.com/giantswarm/microerror"
 	"github.com/giantswarm/micrologger"
-	"k8s.io/api/admission/v1beta1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime"
 
 	"github.com/giantswarm/azure-admission-controller/internal/errors"
 	"github.com/giantswarm/azure-admission-controller/internal/releaseversion"
+	"github.com/giantswarm/azure-admission-controller/pkg/unittest"
 )
 
 var (
@@ -124,13 +122,17 @@ func TestMasterCIDR(t *testing.T) {
 				panic(microerror.JSON(err))
 			}
 
-			admit := &AzureConfigValidator{
-				ctrlClient: fakeCtrlClient,
-				logger:     newLogger,
+			handler, err := NewAzureConfigWebhookHandler(AzureConfigWebhookHandlerConfig{
+				CtrlClient: fakeCtrlClient,
+				Decoder:    unittest.NewFakeDecoder(),
+				Logger:     newLogger,
+			})
+			if err != nil {
+				t.Fatal(err)
 			}
 
-			// Run admission request to validate AzureConfig updates.
-			err = admit.Validate(tc.ctx, getAdmissionRequest(azureConfigRawObj("13.0.0", tc.oldCIDR, tc.oldAZs), azureConfigRawObj("13.0.0", tc.newCIDR, tc.newAZs)))
+			// Run validating webhook handler on AzureConfig update.
+			err = handler.OnUpdateValidate(tc.ctx, azureConfigObj("13.0.0", tc.oldCIDR, tc.oldAZs), azureConfigObj("13.0.0", tc.newCIDR, tc.newAZs))
 
 			// Check if the error is the expected one.
 			switch {
@@ -378,9 +380,13 @@ func TestAzureConfigValidate(t *testing.T) {
 				panic(microerror.JSON(err))
 			}
 
-			admit := &AzureConfigValidator{
-				ctrlClient: fakeCtrlClient,
-				logger:     newLogger,
+			handler, err := NewAzureConfigWebhookHandler(AzureConfigWebhookHandlerConfig{
+				CtrlClient: fakeCtrlClient,
+				Decoder:    unittest.NewFakeDecoder(),
+				Logger:     newLogger,
+			})
+			if err != nil {
+				t.Fatal(err)
 			}
 
 			// Create needed releases.
@@ -389,8 +395,8 @@ func TestAzureConfigValidate(t *testing.T) {
 				t.Fatal(err)
 			}
 
-			// Run admission request to validate AzureConfig updates.
-			err = admit.Validate(tc.ctx, getAdmissionRequest(azureConfigRawObj(tc.oldVersion, "10.0.0.0/24", nil), azureConfigRawObj(tc.newVersion, "10.0.0.0/24", nil)))
+			// Run validating webhook handler on AzureConfig update.
+			err = handler.OnUpdateValidate(tc.ctx, azureConfigObj(tc.oldVersion, "10.0.0.0/24", nil), azureConfigObj(tc.newVersion, "10.0.0.0/24", nil))
 
 			// Check if the error is the expected one.
 			switch {
@@ -407,32 +413,8 @@ func TestAzureConfigValidate(t *testing.T) {
 	}
 }
 
-func getAdmissionRequest(oldRaw []byte, newRaw []byte) *v1beta1.AdmissionRequest {
-	req := &v1beta1.AdmissionRequest{
-		Kind: metav1.GroupVersionKind{
-			Version: "infrastructure.giantswarm.io/v1alpha2",
-			Kind:    "AzureClusterUpgrade",
-		},
-		Resource: metav1.GroupVersionResource{
-			Version:  "provider.giantswarm.io/v1alpha1",
-			Resource: "azureconfigs",
-		},
-		Operation: v1beta1.Update,
-		Object: runtime.RawExtension{
-			Raw:    newRaw,
-			Object: nil,
-		},
-		OldObject: runtime.RawExtension{
-			Raw:    oldRaw,
-			Object: nil,
-		},
-	}
-
-	return req
-}
-
-func azureConfigRawObj(version string, cidr string, azs []int) []byte {
-	azureconfig := providerv1alpha1.AzureConfig{
+func azureConfigObj(version string, cidr string, azs []int) *providerv1alpha1.AzureConfig {
+	azureConfig := providerv1alpha1.AzureConfig{
 		TypeMeta: metav1.TypeMeta{
 			Kind:       "AzureConfig",
 			APIVersion: "provider.giantswarm.io/v1alpha1",
@@ -457,6 +439,6 @@ func azureConfigRawObj(version string, cidr string, azs []int) []byte {
 			VersionBundle: providerv1alpha1.AzureConfigSpecVersionBundle{},
 		},
 	}
-	byt, _ := json.Marshal(azureconfig)
-	return byt
+
+	return &azureConfig
 }