+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - AzureMachine webhook handler that replaces mutators and validators.
 - Spark webhook handler that replaces mutator.
 - AzureClusterConfig webhook handler that replaces validator.
+- AzureConfig webhook handler that replaces validator.
 
 ### Changed